avcodec/bitpacked: add interlace support
This codec is already capable of depacking some combinations of pixel
formats and depth as defined in the RFC4175. The only difference between
progressive and interlace is that either a packet will contain the whole
frame, or only a field of the frame.

As FFmpeg is not capable of handling fields only and recompose an
interlaced frame from that, it has to be done by the codec. To achieve
that, it must use two AVPacket: one for each field (top and bottom).